# services/server/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hot Cross Buns Sync Server API",
    "description": "REST API for the Hot Cross Buns sync server. All endpoints except /health and /api/v1/auth/* require a Bearer JWT token when AUTH_REQUIRED=true.",
    "version": "1.0.0",
    "contact": {
      "name": "Hot Cross Buns"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT with sub=userID, exp=30 days. Not required when AUTH_REQUIRED=false."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "One or more fields are invalid."
              },
              "details": {
                "type": "array",
                "items": { "type": "string" },
                "example": []
              }
            },
            "required": ["code", "message", "details"]
          }
        },
        "required": ["error"]
      },
...
      "MagicLinkRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" }
        },
        "required": ["email"]
      },
      "MagicLinkResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "If that email is registered, a link has been sent." }
        },
        "required": ["message"]
      },
      "VerifyRequest": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "example": "base64url-encoded-token" }
        },
        "required": ["token"]
      },
      "VerifyResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIs..." },
          "expiresAt": { "type": "string", "format": "date-time", "example": "2026-04-21T14:30:00Z" }
        },
        "required": ["token", "expiresAt"]
      },
      "List": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "019513a4-7e2b-7000-8000-000000000001" },
          "userId": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Work" },
          "color": { "type": "string", "nullable": true, "example": "#89b4fa" },
          "sortOrder": { "type": "integer", "example": 0 },
          "isInbox": { "type": "boolean", "example": false },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        },
        "required": ["id", "userId", "name", "sortOrder", "isInbox", "createdAt", "updatedAt"]
      },
      "CreateListRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255, "example": "Work" },
          "color": { "type": "string", "nullable": true, "example": "#89b4fa" }
        },
        "required": ["name"]
      },
      "UpdateListRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "color": { "type": "string", "nullable": true },
          "sortOrder": { "type": "integer" }
        }
      },
...
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/v1/auth/magic-link": {
      "post": {
        "tags": ["auth"],
        "summary": "Request a magic link",
        "description": "Sends a magic-link email to the provided address. Always returns the same response to prevent email enumeration.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MagicLinkRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request accepted (regardless of whether the email exists).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MagicLinkResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify": {
      "post": {
        "tags": ["auth"],
        "summary": "Verify a magic-link token",
        "description": "Validates the magic-link token and returns a JWT session token valid for 30 days.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token verified, JWT issued.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyResponse" }
              }
            }
          },
          "400": {
            "description": "Missing or malformed token.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Invalid or expired magic-link token.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/lists": {
      "post": {
        "tags": ["lists"],
        "summary": "Create a list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateListRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "List created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/List" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["lists"],
        "summary": "Get all lists",
        "description": "Returns all lists for the authenticated user, ordered by sort_order.",
        "responses": {
          "200": {
            "description": "Array of lists.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/List" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/lists/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "List UUID"
        }
      ],
      "get": {
        "tags": ["lists"],
        "summary": "Get a list by ID",
        "responses": {
          "200": {
            "description": "List found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/List" }
              }
            }
          },
          "400": {
            "description": "Invalid ID format.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "List not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["lists"],
        "summary": "Update a list",
        "description": "Partial update. Only include fields to change.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateListRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/List" }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "List not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["lists"],
        "summary": "Delete a list (soft delete)",
        "description": "Soft-deletes the list. Returns 409 if the list is the Inbox.",
        "responses": {
          "204": {
            "description": "List deleted."
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "List not found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "409": {
            "description": "Cannot delete the Inbox list.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}